删除 pod 类型转化为另一种类型，排序后再转换回来，直接使用 pod 方法进行转换

+++ app/http/validator/common/data_type/common_data_type.go
@@ -1,6 +1,7 @@
 package data_type
 
+// 这个其实已经么用了，如果不传入，在后面paginate的时候进行了处理
 type Page struct {
-	Page  float64 `form:"page" json:"page"`
-	Limit float64 `form:"limit" json:"limit"`
+	Page  uint `form:"page" json:"page"`
+	Limit uint `form:"limit" json:"limit"`
 }
